Implemented rudimentary Gibbs sampling for move generation
the "dependent" ant generation is just a stub now and only works because
the ants are ordered by coordinate; ideally, we'd be grouping them more
intelligently around enemy ants.

also, i think move scoring is completely broken.  it seems to behave
mostly right and is a pretty strong bot, but the numbers it outputs make
no sense at all.