huge improvements; gibbs sampling is now a workable solution to the general problem.
each ant's move is conditioned on the moves of one closest ant "above"
and one closest ant "left", so all formations are sampled together in a
chain (and unrelated ants are sampled together but this is usually not
an issue).

also, fixed a major scoring bug.  scoring needs a lot of cleanup.